GML2.0.0/requirements.txt
altgraph==0.17.2
future==0.18.2
ldap3==2.9.1
mpmath==1.2.1
mysql-connector-python==8.0.27
numpy==1.22.0
pefile==2022.5.30
pexpect==4.8.0
protobuf==3.19.1
psutil==5.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
pyqtgraph==0.12.3
PySide2==5.15.2.1
pywin32==303
pywin32-ctypes==0.2.0
PyYAML==6.0
schedule==1.1.0
shiboken2==5.15.2.1
six==1.16.0
WMI==1.5.1
